requirements.txt
ultralytics>=8.0.0
deep-sort-realtime>=1.3.2
opencv-python>=4.8.0
numpy>=1.24.0
streamlit>=1.28.0
paho-mqtt>=1.6.1
